% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_asm.R
\name{calc_asm}
\alias{calc_asm}
\title{Title}
\usage{
calc_asm(sample_list, beta = 0.5, a = 0.2, transform = modulus_sqrt,
  verbose = TRUE)
}
\arguments{
\item{transform}{}

\item{verbose}{}
}
\description{
Title
}

