% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_tracks.R
\name{dame_track_median}
\alias{dame_track_median}
\title{Plot medians or means? per group of score tracks.}
\usage{
dame_track_median(dame, window = 0, positions = 0, derASM = NULL,
  ASM = NULL, colvec = NULL)
}
\arguments{
\item{dame}{GRanges object containing a region of interest, or detected with
find_dames}

\item{window}{Number of CpG sites outside (up or down-stream) of the DAME
should be plotted. Default = 0.}

\item{positions}{Number of bp sites outside (up or down-stream) of the DAME
should be plotted. Default = 0.}

\item{derASM}{SummarizedExperiment object obtained from calc_derivedasm
(Filtering should be done by the user)}

\item{ASM}{SummarizedExperiment object obtained from calc_asm (Filtering
should be done by the user)}

\item{colvec}{Vector of colors (mainly useful for the SNP plot, because I add
it with cowplot, so I don't export a ggplot, optional)}
}
\value{
Plot
}
\description{
Plot medians or means? per group of score tracks.
}
