% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extractbams_output}
\alias{extractbams_output}
\title{extract_bams() output.}
\format{A large list with 8 elements. Each element is a list of
  \code{GRanges} for each sample. Each GRanges in the list includes the
  location of the CpG sites contained in the reads for each SNP. The GRanges
  metadata table contains: 
  \describe{
\item{\code{cov.ref}}{Number of reads of "reference" allele in that SNP}
\item{\code{cov.alt}}{Number of reads of "alternative" allele in that SNP}
\item{\code{meth.ref}}{Number of methylated reads of "reference" allele in
that SNP}
\item{\code{cov.ref}}{Number of methylated reads of "alternative" allele in
that SNP}
\item{\code{snp}}{The SNP containing the reads}
}

For further details, 
see \url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6949/}
sample names in in ArrayExpress do not necessarily match names given here!}
\usage{
extractbams_output
}
\description{
4 Patients from a previous study (Parker et al, 2018.) with colorectal cancer
were sequenced and the normal and cancerous tissue of each patient was
obtained. The data includes a subset of chromosome 19.
}
\keyword{datasets}
