% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tstats.R
\name{get_smoothed_tstats}
\alias{get_smoothed_tstats}
\title{Title}
\usage{
get_smoothed_tstats(betas, Q = 0.9, verbose = TRUE, maxGap = 300)
}
\arguments{
\item{betas}{}

\item{Q}{}

\item{verbose}{}

\item{maxGap}{}
}
\description{
Title
}

