% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tstats.R
\name{get_tstats}
\alias{get_tstats}
\title{Title}
\usage{
get_tstats(sa, design, method = "robust", maxGap = 300, coef = 2,
  verbose = TRUE)
}
\arguments{
\item{sa}{}

\item{method}{}

\item{maxGap}{}

\item{coef}{}
}
\description{
Title
}

