% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_reads.R
\name{getMD}
\alias{getMD}
\title{MDtag parser}
\usage{
getMD(a)
}
\arguments{
\item{a}{Vector of MDtags (single characters)}
}
\value{
A named list of vectors, each vector a parsed version of MDtag:
- nucl.num: Numeric representation of MDtag.
- MDtag: a split version of MDtag
}
\description{
Takes a GenomicAlignments object containing the MDtag, and transforms it
into a vector of characters and numbers
}
\keyword{internal}
