% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bigwig.R
\name{make_bigwig}
\alias{make_bigwig}
\title{Build bigwig files for visualization in IGV}
\usage{
make_bigwig(sample, score.obj, folder, chromsizes.file)
}
\arguments{
\item{sample}{Name of sample (as in colnames from SummarizedExperiment)}

\item{score.obj}{SummarizedExperiment with der.ASM or asm in assays(). Ideally filtered by coverage
previously}

\item{folder}{Destination folder for bigiwig file}

\item{chromsizes.file}{hg19.chrom.sizes.mod or any other chrom.sizes file as in (put link to get this file)}
}
\value{
A bigwig file for the sample chosen.
}
\description{
Builds a bigwig file for each sample, to visualize the ASM score and the derived ASM from
the SummarizedExperiment objects. Requires bedGraphToBigWig to be externally installed.
}
