% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_circle_plot.R
\name{methyl_circle_plot}
\alias{methyl_circle_plot}
\title{Draw methylation circle plot}
\usage{
methyl_circle_plot(snp, vcf.file, bam.file, ref.file, dame = NULL,
  letter.size = 2.5, point.size = 3, sample.name = "sample1",
  cpgsite = NULL)
}
\arguments{
\item{snp}{GRanges object containing SNP location.}

\item{vcf.file}{vcf file.}

\item{bam.file}{bismark bam file path.}

\item{ref.file}{fasta reference file path.}

\item{dame}{(optional) GRanges object containing a region to plot.}

\item{letter.size}{Size of alleles drawn in plot. Default = 2.5.}

\item{point.size}{Size of methylation circles. Default = 3.}

\item{sample.name}{FIX: this is to save the vcf file to not generate it every
time you run the function.}

\item{cpgsite}{(optional) GRanges object containing a single CpG site
location of interest.}
}
\value{
Plot
}
\description{
Draws CpG site methylation status as points, in reads containing a specific
SNP. Generates one plot per bam file.
}
\examples{
DATA_PATH_DIR <- system.file("extdata", ".", package = "DAMEfinder")

get_data_path <- function(file_name) file.path(DATA_PATH_DIR, file_name)
bam_files <- get_data_path("NORM1_chr19_trim.bam")
vcf_files <- get_data_path("NORM1.chr19.trim.vcf")
sample_names <- "NORM1"
reference_file <- get_data_path("19.fa")

snp <- GenomicRanges::GRanges(19, IRanges::IRanges(388065, width = 1))
methyl_circle_plot(snp = snp,
 vcf.file = vcf_files,
 bam.file = bam_files,
 ref.file = reference_file,
 letter.size = 3,
 #dame = dame,
 sample.name = sample_names,
 point.size = 2)

}
