% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_circle_plot.R
\name{methyl_circle_plotCpG}
\alias{methyl_circle_plotCpG}
\title{Draw methylation circle plot without SNP}
\usage{
methyl_circle_plotCpG(cpgsite = cpgsite, bam.file = bam.file,
  point.size = 3, ref.file = ref.file, dame = NULL)
}
\arguments{
\item{cpgsite}{GRanges object containing a single CpG site location of interest}

\item{bam.file}{bismark bam file path}

\item{point.size}{Size of methylation circles. Default = 3.}

\item{ref.file}{fasta reference file path}

\item{dame}{(optional) GRanges object containing a region to plot}
}
\value{
Plot
}
\description{
Draws CpG site methylation status as points, in reads containing a specific CpG site.
Generates one plot per bam file.
}
\examples{

}
