% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_MDS_plot.R
\name{methyl_MDS_plot}
\alias{methyl_MDS_plot}
\title{Multidimensional scaling plot of distances between methylation proportions}
\usage{
methyl_MDS_plot(x, top = 1000, coverage = 4)
}
\arguments{
\item{x}{SummarizedExperiment, output from calc_derivedasm(). TODO: also for calc_asm()?}

\item{top}{Number of top CpG sites used to calculate pairwise distances.}

\item{coverage}{Minimum number of reads covering a CpG site on each allele. Default = 4.
TODO: param colorbygroup TRUE or FALSE. the object should have group informationin colData}
}
\value{
Plot
}
\description{
Same as limma::plotMDS, expect for an arc-sine transformation of the methylation proportions
or scores from 0 to 1.
TODO: check what transformation for ASM-score
}
\examples{


}
