% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_asm.R
\name{modulus_sqrt}
\alias{modulus_sqrt}
\title{Function to calculate signed square root (aka modulus square root)}
\usage{
modulus_sqrt(values)
}
\arguments{
\item{values}{}
}
\description{
Function to calculate signed square root (aka modulus square root)
}

