% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readtuples_output}
\alias{readtuples_output}
\title{read_tuples() output.}
\format{
A large list with 5 elements. Each element is a \code{tibble} with
  the coordinates of the pairs of CpG sites (tuples). Rest of the tibble
  contains:
  \describe{
\item{\code{MM}}{Number of reads with both CpG sites methylated}
\item{\code{MU}}{Number of reads with first CpG site methylated}
\item{\code{UM}}{Number of reads with second CpG site methylated}
\item{\code{UU}}{Number of reads with both CpG sites unmethylated}
\item{\code{cov}}{Coverage, total reads at tuple}
\item{\code{inter_dist}}{Distance in bp between CpG sites}
}

For further details, 
see \url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6949/}
sample names in in ArrayExpress do not necessarily match names given here!
}
\usage{
readtuples_output
}
\description{
3 Patients from a previous study (Parker et al, 2018.) with colorectal cancer
were sequenced and the normal and cancerous tissue of each patient was
obtained. The data includes a subset of chromosome 19. Here one normal sample
is not included.
}
\keyword{datasets}
