% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simes_pval.R
\name{simes_pval}
\alias{simes_pval}
\title{Calculate region-level p-value}
\usage{
simes_pval(sat, sm_tstat, midpt)
}
\arguments{
\item{sat}{Output from \code{\link{get_tstats}}.}

\item{sm_tstat}{(Smoothed) tstat vector from \code{\link{get_tstats}}.}

\item{midpt}{Coordinate vector for each CpG site/tuple.}
}
\value{
Vector of summarized pvals
}
\description{
This function uses the Simes method to calculate a regional-level p-value
based on the single-eBayes p-values. It highly depends on the choice of
\code{maxGap} in \code{find_dames}.
}
\details{
When used as a FDR-control method, for positively correlated P-values, Simes
method is even closer to the nominal alpha level than the Bonferroni-Holm
method.
}
