% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_derivedasm.R
\name{calc_derivedasm}
\alias{calc_derivedasm}
\title{Calculate SNP-based ASM}
\usage{
calc_derivedasm(sample_list, cores = 1, verbose = T)
}
\arguments{
\item{sample_list}{List of samples returned from \code{\link{split_bams}}.}

\item{cores}{Number of cores to thread.}

\item{verbose}{If the function should be verbose.}
}
\value{
\code{RangedSummarizedExperiment} containing in assays:
\itemize{
\item der.ASM: matrix with SNP-based ASM
\item snp.table: Matrix with SNP associated to the CpG site
\item ref.cov: Coverage of the "reference" allele
\item alt.cov: Coevarage of the "alternative" allele
\item ref.meth: Methylated reads from the "reference" allele
\item alt.meth: Methylated reads from the "alternative" allele
}
}
\description{
Combines all the \code{GRangeslist} generated in \code{\link{split_bams}} into a
\code{\link{RangedSummarizedExperiment}} object, and calculates
SNP-based allele-specific methylation.
TODO MAYBE: Make accesor functions to all the assays in this object.
TODO: Doesn't work for one sample
}
\examples{

}
