% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_derivedasm.R
\name{calc_derivedasm}
\alias{calc_derivedasm}
\title{Calculate SNP-based ASM}
\usage{
calc_derivedasm(rds_files, cores = 1, verbose = T)
}
\arguments{
\item{rds_files}{list of .rds files generated by \code{\link{split_bams}}.}

\item{cores}{Number of cores to thread.}

\item{verbose}{If the function should be verbose.}
}
\value{
SummarizedExperiment containing in assays:
\itemize{
\item der.ASM: matrix with SNP-based ASM
\item snp.table: Matrix with SNP associated to the CpG site
\item ref.cov: Coverage of the "reference" allele
\item alt.cov: Coevarage of the "alternative" allele
\item ref.meth: Methylated reads from the "reference" allele
\item alt.meth: Methylated reads from the "alternative" allele
}
}
\description{
Combines all the \emph{GRangeslist} generated in \code{\link{split_bams}} into a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, and calculates
SNP-based allele-specific methylation.
TODO: Make accesor functions to all the assays in this object
}
\examples{

}
