% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_asm.R
\name{calc_score}
\alias{calc_score}
\title{Get a vector of the posterior scores for all positions given beta and a}
\usage{
calc_score(df, beta, a)
}
\arguments{
\item{df}{}
}
\description{
Get a vector of the posterior scores for all positions given beta and a
}

