% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dames.R
\name{find_dames}
\alias{find_dames}
\title{Find DAMEs}
\usage{
find_dames(sa, Q = 0.9, maxGap = 300, verbose = TRUE)
}
\arguments{
\item{sa}{Vector of smoothed t-Statistics obtained from the get_stats function.}

\item{Q}{The percentile set to get a cutoff value K. K is the value on the Qth quantile
of the absolute values of the given smoothed t-Statistics vector. The default is set to 0.9.}

\item{maxGap}{Maximum gap between CpGs in a cluster.}

\item{verbose}{If the function should be verbose.}
}
\value{
A data frame of detected DAMEs ordered by the area each DAME has above the cutoff K.
The larger the area value, the more important the DAME. Each row refers to a DAME and the
following information is provided in the columns:
\itemize{
\item chr: on which chromosome the DAME is found
\item start: The start position of the DAME
\item end: The end position of the DAME
\item value:
\item area: The area of the DAME beyond the exceeded cutoff K.
\item cluster: Then genomic cluster the DAME belongs to. Smoothing of the t-Statistics was
done per cluster.
\item indexStart:
\item indexEnd:
\item L:
\item clusterL:
}
}
\description{
This function finds Differentially Allele-specifically MEthylated regions (DAMEs).
It uses the 'regionFinder' function from the bumphunter package. A region is set as
a DAME if the smoothed t-Statistic, which is set as input to the function, exceeds
a cutoff K. K is set as the value of the 0.9 quantile of the absolute t-Statistics vector.
When the smoothed t-Stats in consecutive CpGs go above K or below -K, the region
is set as a DAME.
}
