% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tstats.R
\name{get_tstats}
\alias{get_tstats}
\title{Get t-Statistics}
\usage{
get_tstats(sa, control, treat, samp.perc = 0.9, coverage = 5,
  method = "ls", maxGap = 20, coef = 2, verbose = TRUE, ...)
}
\arguments{
\item{sa}{A SummarizedExperiment containing ASM values where each row and column correspond to a 
tuple/site and sample respectively.}

\item{control}{Index of columns corresponding to control samples.}

\item{treat}{Index of columns corresponding to treatment samples.}

\item{samp.perc}{At least this percentage of samples per group should have data (no NAs, coverage above threshold)
per row. Default = 0.9.}

\item{coverage}{Minimum number of reads covering a CpG site (sum of both alleles) or tuple. 
Default = 5.}

\item{method}{The method to be used in limma's \code{\link{lmFit}}. The default is set to "ls"
but can also be set to "robust", which is recommended on a real data set.}

\item{maxGap}{The maximum allowed gap between genomic positions for clustering of genomic
regions to be used in smoothing. Default = 20.}

\item{coef}{Columm taken from limma "fit".}

\item{verbose}{Set verbose. Default = TRUE.}

\item{...}{arguments passed to \code{\link{loessByCluster}}}
}
\value{
A vector of smoothed t-Statistics within the \code{RangedSummarizedExperiment}
}
\description{
This function calculates a moderated t-Statistic per site or tuple using \code{limma}'s 
\code{\link{lmFit}} and \code{\link{eBayes}} functions. It then smoothes the obtained 
t-Statistics using \code{bumphunter}'s \code{\link{smoother}} function.
}
\details{
The smoothing is done on genomic clusters consisting of CpGs that are close to each other.
In the case of tuples, the midpoint of the two genomic positions in each tuple is used as the
genomic position of that tuple, to perform the smoothing.The function takes a 
\code{RangedSummarizedExperiment} generated by \code{\link{calc_derivedasm}} or 
\code{\link{calc_asm}} containing ASM across samples, and the index of control and treatment 
samples.
TODO: let user specify more complex design
TODO: I thin I can remove the coef param if above is the case
}
