% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bigwig.R
\name{make_bigwig}
\alias{make_bigwig}
\title{Build bigwig files for visualization in IGV}
\usage{
make_bigwig(sample, scoreObj, folder, chromsizesFile)
}
\arguments{
\item{sample}{Name of sample (as in colnames from
\code{SummarizedExperiment})}

\item{scoreObj}{\code{RangedSummarizedExperiment}, ideally filtered by
coverage from \code{\link{calc_asm}} or \code{\link{calc_derivedasm}}.}

\item{folder}{Destination folder for bigwig file.}

\item{chromsizesFile}{hg19.chrom.sizes or any other file similar specifying
the chromosome sizes.}
}
\value{
A bigwig file for the sample chosen.
}
\description{
Creates a \code{bigwig} file with ASM score for each sample. Requires
\code{bedGraphToBigWig}. installed.
}
\details{
The \code{bedGraphToBigWig} script can be downloaded from
\url{https://github.com/ENCODE-DCC/kentUtils/tree/master/bin/linux.x86_64}.
The \code{chrom.sizes} file can be downloaded from
\url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/bigZips/}, or a similar
goldenPath.
}
\examples{
#To apply to all samples in SummarizedExperiment
#sapply(colnames(ASM_score_matrix), make_bigwig, 
#scoreObj = ASM_score_matrix, folder = , chromsizesFile = )
}
