% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bams.R
\name{split_bams}
\alias{split_bams}
\title{Detect allele-specific methylation from a bam file}
\usage{
split_bams(bam_files, vcf_files, sample_names, reference_file,
  coverage = 2, cores = 1)
}
\arguments{
\item{bam_files}{List of bam files.}

\item{vcf_files}{List of vcf files.}

\item{sample_names}{Names of files in the list.}

\item{reference_file}{fasta file used to generate the bam files.}

\item{coverage}{Minimum number of reads covering a CpG site on each allele. default = 2}

\item{cores}{Number of cores to use. See package {parallel} for description of core. default = 1}
}
\value{
A list of GRanges for each sample. Each list is saved in separate .rds files
}
\description{
The function takes a bam (from bismark) and vcf file for each sample.
For each SNP contained in the vcfile it calculates the proportion of
methylated reads for each CpG site at each allele. At the end it returns
(saves to working directory) a GRanges list, where each GRanges contains
all the CpG sites overlapping the reads containing a specific SNP.
}
